+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## v0.4.4 (pending)
 
+- Add `measures` pipeline
 - Cap Jinja2 version to fix mkdocs
 - Adding the possibility to add context in the processing module
 - Improve the speed of char replacement pipelines (accents and quotes)